[REFACTOR] puthex to another file
