+++ .vscode/launch.json
@@ -1,0 +1,13 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "name": "C/C++ Runner: Debug Session",
+      "type": "lldb",
+      "request": "launch",
+      "args": [],
+      "cwd": "/Users/morishitashoto/Documents/.42/test",
+      "program": "/Users/morishitashoto/Documents/.42/test/build/Debug/outDebug"
+    }
+  ]
+}
