+++ ft_printf.c
@@ -24,20 +24,7 @@ int	ft_printf(const char *str, ...)
 	return (result_length);
 }
 
-int	ft_puthex(unsigned int hex)
-{
-	int	len;
 
-	len = 0;
-	if (hex >= 16)
-	{
-		len += ft_puthex(hex / 16);
-		len += ft_puthex(hex % 16);
-	}
-	else
-		len += ft_putchar("0123456789abcdef"[hex]);
-	return (len);
-}
 
 int	sort_format(va_list *data, const char format)
 {
